Wrap attachment reference lists in inline-uploader widget #929 #942
Add inline-uploader to all fields with type=reference, name=files.

First, the widget is rendered (`inline-uploader-field-template`),
producing a file upload form and a forest for attachment references;
every reference is wrapped in `files-reference-template`, which
produces a view for an attachment instance, which is then rendered
using `file-field-template`.

/uploads server handlers update target instance data, so we re-fetch
it after an upload is finished

fileupload.coffee leftovers are moved to uploads.cofee module which
now contains both #uploads and inline uploads code