db.sh can now read #COMMIT=<commit>/<filename> annotations
If such line occurs somewhere in a patch, database/ snapshot from
<commit> will be checked out instead of the last commit the patch was
changed in. <filename> is mandatory and overrides the patch name.

This is useful when a patch collision resolution requires patch
renaming to avoid version clash. Such a rename will result in a newer
commit being stored for the patch as the last one touching the patch,
thus forcing a newer version of database/ to be checked out when
applying the patch.

Example:

- 1.2.0-patch.sh from a side branch needs database snapshot from
  d34db33f.

- It is renamed to 1.2.1-patch.sh after merging with another branch.
  The rename is commited as b4dc0d3. However, database snapshot from
  b4dc0d3 is too new for 1.2.*-patch.

- Adding a line #COMMIT=d34db33f/1.2.0-patch.sh to the new
  1.2.1-patch.sh forces d34db33f to be used when applying the patch.