Cancel button visibility conditions #216
We now sorta mimick `<field>Reference` observables created for ordinary
reference fields with out `actionsList` observable in the case model.
However, `actionsList` is an observable array, so we can subscribe to
new actions appearing on the screen.

The problem is that actions and services of the case are rendered
asynchronously and cancel button visibility depends on the state of the
case actions.

Odds are that services end up appearing first (before `renderActions` is
first called) and won't be able to subscribe kvm.buttons.cancel.visible
to changes in `actionsList`. To tackle this we set up services hook and
renderActions in a way that ensures that `actionsList` observable is
*created* by whoever reaches it first; `renderActions` pushes actions to
it as they are `modelSetup`'d. Most likely services will create and
subscribe to this observalbe first.

When Case.actions becomes an auto-computed EF field `actionsList` will
need to be created still for the button visibility conditions to work.

Manual dependency handling (and the need to actually define how button
visibility depends on different models rendered on the screen) won't
make implementing #2134 easier.