added old jquery-2.2.4 inside carma #2825
to use it instead of jquery3 with migrate plugins, it has performance
issues, a lot of noise in log and other stuff, it makes sense only if
you're going to work on full migration to remove this migration plugin
before production release. and jquery v2 cannot be added as npm
dependency because is has broken packages for nodejs v8.x.