feature/pure-frontend-getting-rid-of-rxjs:
Got rid of purescript-rxjs dependency.

https://github.com/jasonzoladz/purescript-rxjs package have some issues
affected the CaRMa (such as unsubscribing failure) and maintainer of
that package haven't answered to the issue for few months and it looks
like he is disappeared from github, no activity in 2018 until current
point (1th June 2018) at least.

Anyway, RxJS was used only for "debouncing" changes, no need for whole
thing with a lot of features, "debounce" feature was just implemented by
bare hands and as a result vendor bundle size is reduced.