carma-utils: Carma.Utils.TypeSafe.Generic.Record:
`ConstructorFieldName` and `ConstructorFieldType` now fails on incorrect/unknown
constructor name instead of returning `'Nothing`.

For instance some constraint may check that one of constructors of a data type
does not have a specific field while another one does. That constraint supposed
to return `'Nothing` (when field not presented for in a constructor). But for
any rubbish as a constructor name it also was returning  `'Nothing`, which may
lead to some human-factor mistakes. To avoid that we stating that a constructor
must exists and be correctly written when using those type-families otherwise it
will fail with type-level error.