feature/purescript-0.12.x-migration: done!
It looks like the migration is done, everything compiles and seems to
work. I just did a lot of changes until everything have been compiled
successfully and then I just run it and it seems everything is working
as expected, that is the power of Haskell (well in this particular case
it's about PureScript actually but it doesn't really matter since
they're very close to each other)!